+++ Tests/TuistGeneratorTests/Generator/ProjectGeneratorTests.swift
@@ -51,10 +51,10 @@ final class ProjectGeneratorTests: XCTestCase {
     func test_generate_scheme() throws {
         // Given
         let target = Target.test(name: "Target", platform: .iOS, product: .framework)
-        let scheme = Scheme.test(name: "Target-Scheme", shared: true, buildAction: BuildAction(targets: ["Target"]))
+        let sharedScheme = Scheme.test(name: "Target-Scheme", shared: true, buildAction: BuildAction(targets: ["Target"]))
 
         let targets = [target]
-        let project = Project.test(path: fileHandler.currentPath, name: "Project", targets: targets, schemes: [scheme])
+        let project = Project.test(path: fileHandler.currentPath, name: "Project", targets: targets, schemes: [sharedScheme])
         try fileHandler.touch(fileHandler.currentPath.appending(component: "Project.swift"))
         
         let cache = GraphLoaderCache()
@@ -74,6 +74,35 @@ final class ProjectGeneratorTests: XCTestCase {
         let schemesPath = got.path.appending(RelativePath("xcshareddata/xcschemes"))
         let targetScheme = schemesPath.appending(component: "Target-Scheme.xcscheme")
         XCTAssertTrue(fileHandler.exists(targetScheme))
+    }
+
+    func test_generate_local_scheme() throws {
+        // Given
+        let target = Target.test(name: "Target", platform: .iOS, product: .framework)
+        let localScheme = Scheme.test(name: "Target-Local", shared: false, buildAction: BuildAction(targets: ["Target"]))
+        
+        let targets = [target]
+        let project = Project.test(path: fileHandler.currentPath, name: "Project", targets: targets, schemes: [localScheme])
+        try fileHandler.touch(fileHandler.currentPath.appending(component: "Project.swift"))
+        
+        let cache = GraphLoaderCache()
+        cache.add(project: project)
+        let graph = Graph.test(entryPath: fileHandler.currentPath,
+                               cache: cache,
+                               entryNodes: [TargetNode(project: project,
+                                                       target: target,
+                                                       dependencies: [])])
+        
+        // When
+        let got = try subject.generate(project: project,
+                                       options: GenerationOptions(),
+                                       graph: graph)
+        
+        // Then
+        let username = NSUserName()
+        let userSchemesPath = got.path.appending(RelativePath("xcuserdata/\(username).xcuserdatad/xcschemes"))
+        let userScheme = userSchemesPath.appending(component: "Target-Local.xcscheme")
+        XCTAssertTrue(fileHandler.exists(userScheme))
     }
 
     func test_generate_testTargetIdentity() throws {